# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image Cleaning",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/uploads": {
      "post": {
        "summary": "Reserve",
        "operationId": "reserve_v1_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Upload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/assets": {
      "get": {
        "summary": "Assets",
        "operationId": "assets_v1_assets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_v1_jobs_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Jobs",
        "operationId": "jobs_v1_jobs_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/searches": {
      "post": {
        "summary": "Search",
        "operationId": "search_v1_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Box": {
        "properties": {
          "x_min": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X Min"
          },
          "y_min": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y Min"
          },
          "x_max": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X Max"
          },
          "y_max": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y Max"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x_min",
          "y_min",
          "x_max",
          "y_max"
        ],
        "title": "Box"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Inpaint": {
        "properties": {
          "erase_mask_asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Erase Mask Asset Id"
          },
          "method": {
            "type": "string",
            "const": "classical",
            "title": "Method",
            "default": "classical"
          },
          "feather_px": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Feather Px",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "erase_mask_asset_id"
        ],
        "title": "Inpaint"
      },
      "JobRequest": {
        "properties": {
          "input_asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Input Asset Id"
          },
          "segmentation": {
            "$ref": "#/components/schemas/Segmentation"
          },
          "restoration": {
            "$ref": "#/components/schemas/Restoration"
          },
          "upscale": {
            "$ref": "#/components/schemas/Upscale"
          },
          "output": {
            "$ref": "#/components/schemas/Output"
          },
          "memory": {
            "$ref": "#/components/schemas/Memory"
          },
          "inpaint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Inpaint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input_asset_id"
        ],
        "title": "JobRequest"
      },
      "Memory": {
        "properties": {
          "allow_indexing": {
            "type": "boolean",
            "title": "Allow Indexing",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Memory"
      },
      "Output": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "png",
              "jpeg"
            ],
            "title": "Format",
            "default": "png"
          },
          "background": {
            "type": "string",
            "title": "Background",
            "default": "transparent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Output"
      },
      "Point": {
        "properties": {
          "x": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y"
          },
          "label": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "label"
        ],
        "title": "Point"
      },
      "Restoration": {
        "properties": {
          "denoise": {
            "type": "string",
            "enum": [
              "off",
              "classical"
            ],
            "title": "Denoise",
            "default": "off"
          },
          "deblur": {
            "type": "string",
            "const": "off",
            "title": "Deblur",
            "default": "off"
          },
          "tone": {
            "type": "string",
            "enum": [
              "off",
              "classical"
            ],
            "title": "Tone",
            "default": "off"
          },
          "strength": {
            "type": "number",
            "maximum": 0.5,
            "minimum": 0.0,
            "title": "Strength",
            "default": 0.2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Restoration"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 24
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "Segmentation": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "automatic",
              "interactive",
              "none"
            ],
            "title": "Mode",
            "default": "automatic"
          },
          "allow_lightweight_fallback": {
            "type": "boolean",
            "title": "Allow Lightweight Fallback",
            "default": false
          },
          "coordinate_space": {
            "type": "string",
            "const": "canonical_normalized",
            "title": "Coordinate Space",
            "default": "canonical_normalized"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Points"
          },
          "box": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Box"
              },
              {
                "type": "null"
              }
            ]
          },
          "refinement_mask_asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refinement Mask Asset Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Segmentation"
      },
      "Upload": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 200,
            "title": "Filename",
            "default": "image"
          },
          "mime": {
            "type": "string",
            "enum": [
              "image/png",
              "image/jpeg",
              "image/webp"
            ],
            "title": "Mime"
          },
          "bytes": {
            "type": "integer",
            "maximum": 26214400.0,
            "exclusiveMinimum": 0.0,
            "title": "Bytes"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256"
          },
          "role": {
            "type": "string",
            "enum": [
              "original",
              "refinement_mask",
              "erase_mask"
            ],
            "title": "Role",
            "default": "original"
          },
          "parent_asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Asset Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mime",
          "bytes",
          "sha256"
        ],
        "title": "Upload"
      },
      "Upscale": {
        "properties": {
          "scale": {
            "type": "integer",
            "enum": [
              1,
              2
            ],
            "title": "Scale",
            "default": 1
          },
          "method": {
            "type": "string",
            "const": "lanczos",
            "title": "Method",
            "default": "lanczos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Upscale"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}